# src/static/swagger.yaml
openapi: 3.0.1
info:
  title: Events API
  description: >-
    GitHub integration API is a service to connect to the GitHub with personal token or GitHub app.
  contact:
    email: admin@mail.com
  version: 1.0.0
externalDocs:
  description: Find out more about Swagger
  url: 'http://swagger.io'
servers:
  - url: 'https://events-api-nqc53wwloq-ez.a.run.app/'
  - url: 'http://events-api-nqc53wwloq-ez.a.run.app/'
tags:
  - name: oauth
    description: Everything related to OAUth
paths:
  /oauth:
    get:
      tags:
        - oauth
      description: The OAuth proccess connect to GitHub
      summary: Performs the OAuth process step one
      responses:
        '302':
          description: Redirect to service's authorization page
  /oauth/github-callback:
    get:
      tags:
        - oauth
      description: The OAuth process callback from github
      summary: Perform the OAuth process callback from github
      responses:
        '302':
          description: Redirects to a home page on success
        '400':
          description: On authorization failure
          content:
            application/json:
              schema:
                type: object
                properties:
                  invalid_field_name:
                    type: string
                    example: error description
  /oauth/access-token:
    post:
      tags:
        - personal_token
      description: Connect with personal token
      summary: Connect with personal token
      security:
        - Token: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  personal_token:
                    type: string
                    example: f902b0fc892b3c0b1ae4aaab45b8d32f
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
components:
  responses:
    UnauthorizedResponse:
      description: Token authentication failed
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: authentication failure detailed info
  securitySchemes:
    Token:
      type: apiKey
      in: header
      name: Authorization